Update and rename register.js to Register.js
